Mission Mode Override Command

+++ src/MissionMode.hpp
@@ -16,7 +16,7 @@ public:
     {
         start_time = new_start_time;
     }
-    virtual int get_id();
+    virtual uint8_t get_id();
     virtual std::string get_name();
     virtual mode_type get_type();
 };
@@ -26,7 +26,7 @@ class Boot : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 0;
     }
@@ -45,7 +45,7 @@ class AliveSignal : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 1;
     }
@@ -64,7 +64,7 @@ class DetumbleSpin : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 2;
     }
@@ -83,7 +83,7 @@ class LowPowerDetumbleSpin : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 3;
     }
@@ -102,7 +102,7 @@ class Normal : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 4;
     }
@@ -121,7 +121,7 @@ class Transmit : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 5;
     }
@@ -140,7 +140,7 @@ class LowPower : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 6;
     }
@@ -159,7 +159,7 @@ class NormalDeployment : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 7;
     }
@@ -178,7 +178,7 @@ class TransmitDeployment : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 8;
     }
@@ -197,7 +197,7 @@ class LowPowerDeployment : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 9;
     }
@@ -216,7 +216,7 @@ class NormalArmed : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 10;
     }
@@ -235,7 +235,7 @@ class TransmitArmed : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 11;
     }
@@ -254,7 +254,7 @@ class LowPowerArmed : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 12;
     }
@@ -273,7 +273,7 @@ class NormalInSun : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 13;
     }
@@ -292,7 +292,7 @@ class TransmitInSun : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 14;
     }
@@ -311,7 +311,7 @@ class LowPowerInSun : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 15;
     }
@@ -330,7 +330,7 @@ class VoltageFailureInSun : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 16;
     }
@@ -349,7 +349,7 @@ class BootIMU : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 17;
     }
@@ -368,7 +368,7 @@ class BootCamera : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 18;
     }
@@ -387,7 +387,7 @@ class MandatoryBurns : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 19;
     }
@@ -406,7 +406,7 @@ class RegularBurns : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 20;
     }
@@ -425,7 +425,7 @@ class Photo : public MissionMode
 public:
     void transition_to();
     void dispatch();
-    int get_id()
+    uint8_t get_id()
     {
         return 21;
     }
